polarization algorithms: usable for multigraphs
Before this commit, sometimes edges were accessed with `G.edges()` and
sometimes they were accessed with `G.edges`. This commit replaces the
later by the former.

The advantage of the former is that is also usable for multigraphs. When
using graphs, both methods return a tuple `(s, t)`. But when using
multigraphs, `G.edges` returns a tuple `(s, t, k)` with a key `k`, while
`G.edges()` returns only a tuple `(s, t)`, as expected by the rest of the
code.